Report/решения/2.LLM_Generated_Code/AnalyzeAce/requirements.txt
openai
pyyaml
polars
numpy
mathplotlib
seaborn
gradio